+++ include/core/commands.h
@@ -54,6 +54,9 @@ void handle_webuiap_cmd(int argc, char **argv);
 // BadUSB and USB keyboard host
 void handle_badusb_cmd(int argc, char **argv);
 void handle_usb_kbd_cmd(int argc, char **argv);
+#ifdef CONFIG_HAS_BADBLE
+void handle_badble_cmd(int argc, char **argv);
+#endif
 
 #if CONFIG_ENABLE_GHOSTSCRIPT
 // GhostScript